requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
greenlet==1.1.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
tqdm==4.61.1
urllib3==1.26.5
python-dotenv
psycopg2-binary